✨ Dummy for installation details and errors
- New entry point for dummy data source
- Erp data sources exceptions separated into a file to be imported by
  dummy
- installation list includes entries to make details fail
