+++ backend/datasources/erp.py
@@ -6,52 +6,21 @@ from ..models import TokenUser, UserProfile, SignatureResult, InstallationSummar
 from .. import erp
 from ..utils.gravatar import gravatar
 from ..utils.vat import nif2vat
-
-class ErpError(Exception):
-    def __init__(self, erp_error: dict):
-        formatted_trace = ''.join(erp_error.get('trace', "No trace available"))
-        super().__init__((
-            "{code}\n"
-            "{error}\n\n"
-            "Remote backtrace:\n {formatted_trace}"
-        ).format(
-            formatted_trace=formatted_trace,
-            **erp_error,
-        ))
-
-class ErpValidationError(ErpError):
-    def __init__(self, exception: ValidationError):
-        message = (
-            "Invalid data received from ERP\n"
-            f"{ns(error=ns.loads(exception.json())).dump()}"
-        )
-        error(message)
-        self.original = exception
-        super().__init__(dict(
-            code="ErpValidationError",
-            error=message,
-        ))
+from .exceptions import(
+    ErpError,
+    ErpValidationError,
+    ContractWithoutInstallation,
+    ContractNotExists,
+    UnauthorizedAccess,
+    NoSuchUser,
+    NoDocumentVersions,
+)
 
 @contextmanager
 def catchValidationErrors():
     try: yield
     except ValidationError as exception:
         raise ErpValidationError(exception)
-
-class ContractWithoutInstallation(ErpError):
-    pass
-
-class ContractNotExists(ErpError):
-    pass
-
-class UnauthorizedAccess(ErpError):
-    pass
-
-class NoSuchUser(ErpError):
-    pass
-
-class NoDocumentVersions(ErpError):
-    pass
 
 expected_erp_exceptions = [
     ContractWithoutInstallation,
